refactor module path extract
