+++ ScriperSol/ScriperLib/Runners/PythonRunner.cs
@@ -1,4 +1,5 @@
 ﻿using IronPython.Hosting;
+using Microsoft.Scripting.Hosting;
 using ScriperLib.Enums;
 using ScriperLib.Extensions;
 using ScriperLib.Outputs;
@@ -6,8 +7,6 @@ using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Threading.Tasks;
-using Microsoft.CodeAnalysis.Operations;
-using Microsoft.Scripting.Hosting;
 
 namespace ScriperLib.Runners
 {
@@ -54,35 +53,10 @@ namespace ScriperLib.Runners
             var paths = engine.GetSearchPaths();
             paths.Add(Path.GetDirectoryName(script.Configuration.Path));
 
-            const string importStart = "#${";
-
-            using var reader = new StreamReader(script.Configuration.Path);
-            while (true)
+            var scriptPaths = ModulePathExtractor.ExtractPaths(script.Configuration.Path);
+            foreach (var path in scriptPaths)
             {
-                var line = reader.ReadLine();
-
-                if (line is null)
-                {
-                    break;
-                }
-                if (string.IsNullOrEmpty(line))
-                {
-                    continue;
-                }
-
-                if (line.Contains(importStart))
-                {
-                    var start = line.IndexOf(importStart);
-                    var end = line.IndexOf("}");
-
-                    if (end < start)
-                    {
-                        //throw
-                    }
-
-                    var path = line.Substring(start + importStart.Length, end - (start+ importStart.Length));
-                    paths.Add(path);
-                }
+                paths.Add(path);   
             }
 
             engine.SetSearchPaths(paths);
